Create Product List page
I've functionality for filtering the four required properties,
name, brand, store, and category,
with the first two being simple inputs,
and the second two being selectors.
For now, I left store as a choice between
Willies and PdT to reflect Susan's grocery shopping.

I've also tried an implementation of a
MatPaginator, but need to fork over the server-side of Products
before I can check to see if I built it right.