Fix displaying add product issues
Fixed the issue where add product was not being displayed properly. Adding a product also does not work, it continues to give the error for not being able to create a product. It also seems to be tampering with the ability to properly filter products on the product list page.