Clean up product-list cypress tests
There were too many "tests" and although the assertions were useful, grouping them separately was not. I made the testing for opening and closing extension panels appear as one test with many assertions, so that when tests for delete are added, there will not be 5 tests for each category. That would result in 50 tests for the product dropdowns alone.