Debug server-side product card bad request issue
There was a bad request when the product was clicked and it was supposed to take you to the product card on the single product page screen. It went to a blank screen, so I altered some of the server routes. Change product had the same route as getProductId, despite having different uses. So I changed it to the way that Mental Pulp Sensation (MPS) routes worked. The single product page is no longer blank.