Made progress with editing
Got it to edit just the threshold field without autofilling + it will change if nothing is inputed. We did figure out that all the connections work, however. So we are in a spot where we need to figure out how to get the formBuilder to autofill for us without blowing up.
Co-authored-by: Kahnurd <hetti031@morris.umn.edu>