Convert categories to field names
There are a lot places where you have duplicate code when you could be looping over categories.

This changes how `ProductCategory` is defined so that we can enumerate over categories, as suggested here: https://stackoverflow.com/a/64174790

The new function `categoryCamelCase()` converts categories into a camel case form with `Items` appended at the end, i.e., `'pet supplies'` is mapped to `petSuppliesItems`.

I *think* you'd eventually want to get rid of this function, perhaps by either just doing everything with the camel case version, or by having an enum that maps the camel case version to the version with the spaces.

I'm also not sure it you want to be adding the suffix `Items` here in this function. It was needed everywhere I called `categoryCamelCase` so I put it here, but I feel like there might be places where you want the camel case version but don't want/need 'Items' at the end. If that's true, then we probably want to move adding 'Items' off to where it's actually needed.