Use `categoryCamelCase` to simplify a lot of code
This illustrates how we can use the new `categoryCameCase()` function to eliminate a lot of duplicate code in `PantryProductsList`.

Both of these simplifications rely on the ability to access fields in JavaScript with `[]` as well as the dot notation. So, for example, `x.y` can be replaced with `x['y']` or `x[s]` where `s = 'y'`. This allows us to compute field names (by calling `categoryCamelCase()`) and then access them using `[]`.

We first replace the huge list of categorization calls in `makeCategoryLists()` into a short `forEach` loop.

We could replace the big list in `removeItem` with a loop like above. We didn't need to do that, though, because a `PantryItem` can be in only one category. So we can use `categoryCamelCase(item.category)` to compute the necessary field name, and then call filter *just once* on the necessary category instead of on _all_ the categories.