Create validation form method between add/edit pages + cypress tests
Made a static public method that outputs all of the Validation strings we want for form builder. Both edit and add use this method now to cut down on repeated code. The formbuilder is still built the same way, so it would be useful to make sure they are always the same still. 
Co-authored-by: Kahnurd <hetti031@morris.umn.edu>